requirements.txt
# Web Framework & Validation
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5

# Redis & Async
redis==5.0.1

# Security & Authentication
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.12.0
python-multipart==0.0.22
